# backend/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PrimeTrade SignalHub API Documentation",
    "version": "1.0.0",
    "description": "Production-grade API documentation for PrimeTrade SignalHub - a high-scalability crypto trading signals management platform."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    }
  ],
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "example": "John Doe" },
                  "email": { "type": "string", "format": "email", "example": "john@primetrade.com" },
                  "password": { "type": "string", "minLength": 6, "example": "UserPass123!" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "User registered successfully." },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "string", "format": "uuid" },
                            "name": { "type": "string" },
                            "email": { "type": "string" },
                            "role": { "type": "string", "example": "USER" },
                            "createdAt": { "type": "string", "format": "date-time" },
                            "updatedAt": { "type": "string", "format": "date-time" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation failed" },
          "409": { "description": "Email already registered" }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Log in an existing user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "john@primetrade.com" },
                  "password": { "type": "string", "example": "UserPass123!" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Login successful." },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "string", "format": "uuid" },
                            "name": { "type": "string" },
                            "email": { "type": "string" },
                            "role": { "type": "string", "example": "USER" }
                          }
                        },
                        "accessToken": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsIn..." },
                        "refreshToken": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsIn..." }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh an expired access token using refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsIn..." }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens rotated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Tokens refreshed successfully." },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": { "type": "string" },
                        "refreshToken": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid or expired session" }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Log out and invalidate session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsIn..." }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        }
      }
    },
    "/api/v1/signals": {
      "get": {
        "tags": ["Signals"],
        "summary": "Retrieve all trading signals belonging to the authenticated user",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of signals retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "signals": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Signal" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthenticated" }
        }
      },
      "post": {
        "tags": ["Signals"],
        "summary": "Create a new trading signal",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["assetSymbol", "signalType", "entryPrice", "targetPrice", "stopLoss"],
                "properties": {
                  "assetSymbol": { "type": "string", "example": "BTC" },
                  "signalType": { "type": "string", "enum": ["BUY", "SELL"], "example": "BUY" },
                  "entryPrice": { "type": "number", "example": 67500 },
                  "targetPrice": { "type": "number", "example": 72000 },
                  "stopLoss": { "type": "number", "example": 65000 },
                  "notes": { "type": "string", "example": "Support breakout." }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signal created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Signal created successfully." },
                    "data": {
                      "type": "object",
                      "properties": {
                        "signal": { "$ref": "#/components/schemas/Signal" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthenticated" }
        }
      }
    },
    "/api/v1/signals/{id}": {
      "get": {
        "tags": ["Signals"],
        "summary": "Retrieve details of a single signal",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Signal UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Signal details returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "signal": { "$ref": "#/components/schemas/Signal" }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": { "description": "Access forbidden" },
          "404": { "description": "Signal not found" }
        }
      },
      "put": {
        "tags": ["Signals"],
        "summary": "Update a trading signal",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Signal UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "entryPrice": { "type": "number" },
                  "targetPrice": { "type": "number" },
                  "stopLoss": { "type": "number" },
                  "notes": { "type": "string" },
                  "status": { "type": "string", "enum": ["OPEN", "CLOSED"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signal updated successfully"
          },
          "403": { "description": "Access forbidden" },
          "404": { "description": "Signal not found" }
        }
      },
      "delete": {
        "tags": ["Signals"],
        "summary": "Delete a trading signal",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Signal UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Signal deleted successfully"
          },
          "403": { "description": "Access forbidden" },
          "404": { "description": "Signal not found" }
        }
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "tags": ["Admin Operations"],
        "summary": "Get list of all platform users (Admin only)",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Successful retrieval of users directory"
          },
          "403": { "description": "Forbidden - requires Admin role" }
        }
      }
    },
    "/api/v1/admin/audit-logs": {
      "get": {
        "tags": ["Admin Operations"],
        "summary": "Get list of all database audit logs (Admin only)",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Successful retrieval of audit log directory"
          },
          "403": { "description": "Forbidden - requires Admin role" }
        }
      }
    },
    "/api/v1/market/price/{symbol}": {
      "get": {
        "tags": ["Market Data"],
        "summary": "Get the live market price for a cryptocurrency",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "description": "Ticker symbol (e.g. BTC, ETH)",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with live price",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "symbol": { "type": "string", "example": "BTC" },
                        "price": { "type": "number", "example": 68432.12 },
                        "currency": { "type": "string", "example": "USD" },
                        "source": { "type": "string", "example": "Binance API" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid symbol format" },
          "502": { "description": "API Gateway error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Signal": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "assetSymbol": { "type": "string" },
          "signalType": { "type": "string", "enum": ["BUY", "SELL"] },
          "entryPrice": { "type": "number" },
          "targetPrice": { "type": "number" },
          "stopLoss": { "type": "number" },
          "status": { "type": "string", "enum": ["OPEN", "CLOSED"] },
          "notes": { "type": "string", "nullable": true },
          "userId": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}